Initial commit.
This commit copies over the code from Squarespace's internal pgbedrock
repo to this public repo. It also adds the ability to push versioned
releases to quay (for docker) and pypi (for Python). Note that the pypi
release process is untested as we need a version to be ready first
before we can push.