Change package version management to use .py file
Previously we tried to manage package version in a simple file at the
root of the repo. This caused trouble at the interface of docker,
doing `pip install .`, and using a console entrypoint. For the moment
it's easier just to put the version in the __init__.py and use grep /
regex to parse the version out for our Makefile and setup.py commands.