Fix improper schema in several tests
pgbedrock expects a schema to be a dict of role definitions, each of
which is a dict. A previous (pre-open-source) version of pgbedrock
instead had a list of role definitions, and a few tests still reflected
that assumption. While these tests pass because the
core_configure.load_spec() function will accept that input, it is
confusing as developers may think this is a reasonable spec definition.
For clarity / consistency's sake this commit switches these test specs
to use a schema that is actually valid, i.e. a dict of role definitions.