Remove sole reference to pg_roles
We had 1 reference to pg_roles. Since we are supporting multiple
Postgres versions and eventually possibly across other implementations
(e.g. Redshift et al.) it makes sense to consolidate on the smallest
number of operational tables possible in order to limit the number of
things that may be missing / different across versions.

As we use pg_authid everywhere else and require it for password
validation, this commit switches the 1 pg_roles reference over to
using pg_authid as well.