Extract fetch object attributes query as new method
In the interest of removing duplicative queries and ensuring consistent
data across queries, we will be combining the queries that back:
    * DatabaseContext.get_all_nonschema_objects_and_owners()
    * DatabaseContext.get_all_object_attributes()

These two methods will use the same backing query, and while we could
just have method A fetch the data and method B just call method A and
transform its results, that adds more Python logic and more chances to
screw things up. It is cleaner to just run that query, cache the results
of the method call, and then let each of those methods use those raw
query results.

This commit does the first part of that last idea: it extracts our the
common query and uses it in the method that previously fetched the data
itself. A subsequent commit will switch the other method to use this
pre-fetched data as well.