implement travis-ci
We are doing some extra work sort of re-implementing how the makefile
sets up your test environment locally, but this is because the makefile
assumes you can provision docker containers and setup portforwarding of
the running containers so that you can test pgbedrock totally under
docker or use a mix of docker and your installed python version.

Travis prevents us from using the makefile's current setup verbatim
because we can't do the same port forwarding, so we either can re-setup
the environment using what Travis provides for us, or awkwardly create a
bunch of Travis-aware makefile targets.  I feel like the build isn't
complicated enough to warrant a TON of reuse between the different ways.
Asking Travis for a Postgres db and removing a couple roles isn't so
bad.