Rename method to follow DatabaseContext.get_* format
Originally this method was named 'fetch_*' to indicate that it will be
doing network activity. The plan was to rename all methods that actually
run queries to make this distinction clearer. However, these methods
also cache their results, so most calls actually are very fast and do
not have network latency. As a result, naming these methods with a
convention to indicate more work is done (i.e. fetch_*) is misleading.
This commit goes back to using get_* for the one method that had been
written to use fetch_*.