Extract spec inspection and loading into submodule
We have several hundred lines of spec-related code, most of which does
validation of various conditions. As we will be adding more validation
soon (e.g. when we add table and sequence support) and more parsing
logic soon (e.g. when we add multi-database support), it makes sense to
extract this functionality into a submodule so we don't clutter the
core_configure module and we keep related functionality close to each
other.