Move schema test functionality into spec_inspector
Previously we had some functionality within attributes.py and
ownerships.py to assert characteristics of a spec before running each of
those submodules, failing and throwing an exception if those conditions
weren't met.

This commit moves those verification functions into the
spec_inspector.py module. This is done for a couple reasons:

    1) Fail early. It sucks to have pgbedrock run for a bit and then
       fail halfway through. It'd be nice to fail immediately so users
       can iterate faster.
    2) Combine related functionality. This is more of a stylistic
       choice. We could put these verification functions close to what
       they're verifying, but we're also going to be doing a bunch of
       verification regardless of the submodule (e.g. make sure there
       aren't duplicate roles, since that could screw everything up),
       so it seems as reasonable to put all of this together, which
       makes it easier for a dev to evaluate all of the assertions we
       are checking against specs.

In addition to the above, this commit renames some of the verification
methods to make the names more consistent and clearer.