Support testing multiple Postgres versions
We currently only test against Postgres 9.6.4, but our users may be
using different versions of Postgres. While it isn't clear yet how far
back we're willing to support, it's reasonable to support at least one
version behind so we'll want to be able to support multiple versions,
regardless what those versions may be.

This commit adds that support for Postgres 9.6.4 and Postgres 9.5.13,
but stops one step short of adding support for Postgres 10 because that
causes some tests to fail. A subsequent commit will add Postgres 10
support.