Formally support Postgres 10
Prior to this commit Postgres 10 still worked with pgbedrock, but
pgbedrock's test suite would fail with a Postgres 10.x docker container.
This was due to new roles that were introduced by Postgres 10 which the
test infrastructure did not know to expect. As a result, those roles
caused test failures.

This commit updates the test infrastructure to properly account for these
new roles, allowing us to formally support Postgres 10 from here on out.