Have DBObject.qualified_name be foo.* not foo."*"
Doing schema."table" makes sense for real objects, but * isn't a real
object. As a result, doing foo."*" is confusing: it seems like we're
talking about a real object name '*', but we're not. To avoid this
confusion, we output foo."bar" for everything except when the
object_name is a *, in which case we do foo.*