Rename ObjectName.object_name to .unqualified_name
ObjectName.object_name is confusing (it should be redundant but isn't:
we're referring to the object name without a schema). Unfortunately,
Postgres refers to entities as "objects", so we don't want to just
called things ObjectName.object either. As a result, the clearest name
is probably just ObjectName.unqualified_name, so we switch to that.