Create way to convert spec strings to ObjectNames
All lists within the owns and privileges subdicts will be converted to
ObjectNames. If there are any empty entries, e.g.:

    roleA:
        owns:
            tables:

    ...

Then these empty entries get converted to None by PyYAML. Rather than
override this, I'm leaving this behavior as:
    1) It adds a some complexity to try to guard against Nones in a
    bunch of spots.
    2) More than the complexity, it makes the code less readable
    3) And most importantly, the point of pgbedrock is to make a spec
    simply to read and follow. Those empty entries are unnecessary
    cruft.

As a result, it is both easier on our end and more correct based on this
tool's philosophical purpose to not try to convert entries from None to
an empty list.

Originally, I had tried to use a customer PyYAML constructor to
automatically convert these sublists from strings to ObjectName
instances, but I found PyYAML confusing and complicated enough that
whoever would have to maintain this code if it breaks would probably end
up quite confused. Because of that it seemed wiser to just let PyYAML do
its thing and then we come in with regular Python and convert the
relevant parts of the loaded spec into ObjectName instances.