Correct schema name quoting
Currently, certain `ALTER... SCHEMA` statements will fail if they address a
schema whose name contains a hyphen, which can happen, for instance, managing
the personal schema of a user whose name contains a hyphen. This commit wraps
the schema name in quotes for the two affected sql queries, and bumps pgbedrock
to 0.3.2.

Closes #43.