Add a test to cover more complex dash-in-role-name case
There exists a test for a role with a dash in it having a personal schema.
However, the test case is too simple, and does not force pgbedrock to configure
default permissions. This test case *does* force default permission
configuration, so we can see that roles with dashes and personal schemas is, in
fact, broken.